Launch rocket and display distance and speed.

+++ cs/game1/Engineer.cs
@@ -2,6 +2,6 @@ namespace Game1
 {
     class Engineer
     {
-        public int SkillLevel { set; get; } = 0.3f;
+        public float SkillLevel { set; get; } = 0.3f;
     }
 }
